+++ NEWS.md
@@ -1,6 +1,6 @@
 # isstatic development
 
-* Add `is_file()` and `is_dir()` functions.
+* Add `is_file()`, `is_dir()`, `has_min_length()` and `has_max_length()` functions.
 * Make `str_fileext()` length stable (although functions may now return NA values).
 
 # isstatic 0.1.0.9000 (2023-03-15)